+++ plugins/provisioners/chef/config/chef_client.rb
@@ -11,7 +11,6 @@ module VagrantPlugins
         attr_accessor :delete_client
         attr_accessor :delete_node
         attr_accessor :encrypted_data_bag_secret_key_path
-        attr_accessor :encrypted_data_bag_secret
         attr_accessor :environment
         attr_accessor :validation_key_path
         attr_accessor :validation_client_name
@@ -24,10 +23,15 @@ module VagrantPlugins
           @delete_client                      = UNSET_VALUE
           @delete_node                        = UNSET_VALUE
           @encrypted_data_bag_secret_key_path = UNSET_VALUE
-          @encrypted_data_bag_secret          = UNSET_VALUE
           @environment                        = UNSET_VALUE
           @validation_key_path                = UNSET_VALUE
           @validation_client_name             = UNSET_VALUE
+        end
+
+        def encrypted_data_bag_secret=(value)
+          puts "DEPRECATION: Chef encrypted_data_bag_secret has no effect anymore."
+          puts "Remove this from your Vagrantfile since it'll be removed in the next"
+          puts "Vagrant version."
         end
 
         def finalize!
@@ -38,7 +42,6 @@ module VagrantPlugins
           @delete_client = false if @delete_client == UNSET_VALUE
           @delete_node = false if @delete_node == UNSET_VALUE
           @encrypted_data_bag_secret_key_path = nil if @encrypted_data_bag_secret_key_path == UNSET_VALUE
-          @encrypted_data_bag_secret          = "/tmp/encrypted_data_bag_secret" if @encrypted_data_bag_secret == UNSET_VALUE
           @environment = nil if @environment == UNSET_VALUE
           @validation_client_name = "chef-validator" if @validation_client_name == UNSET_VALUE
           @validation_key_path = nil if @validation_key_path == UNSET_VALUE